restore Redundant renaming
